moved camera component to components folder
